+++ HISTORY.rst
@@ -10,6 +10,8 @@ History
 * statemachine integrity checks are performed at import time instead of when
   instance creation. This allows early feedback on invalid definitions.
 * Drop support for adding `validators` as assigment to `Event`.
+* `TransitionNotAllowed` changed internal attr from `transition` to `event`.
+* `StateMachine.run` is deprecated in favor of `StateMachine.send`.
 
 
 0.9.0 (2022-12-21)